+++ CHANGELOG.md
@@ -25,6 +25,7 @@ TODO
 - Selectors: PingLatency, ReversePath, ...
 - Fix PplPolicy.fromJson()
 - Add PPL JSON+YAML export. Fix JSON import of multiple policies
+- Authenticate paths!?!?!?
 
 - We could also do revers-lookup inside service.lookup() -> e.g. works for "129.132.175.104"
   -> is a SCION enabled IP; this gives us the ISD/AS.
